Put ecma parsing test into a zero-separated file
This reduces the size of literal strings which make some compilers whine.

Also enable all tests, including labels and for..in which are now fully implemented
(and were already covered in manual tests).

Now parses 11112 ecma262 tests. The remaining 454 tests are negative tests,
which we should eventually check as well.