Use the nice `test.mk` we have in fossa
In fossa we gradually ended up with a few nice things in the
makefile, let's reuse them here:

 - ASAN
 - GCOV and LCOV
 - multiple dialects
 - quick parallel dialect syntax check

This change is a little bit rough, since some stuff remains in the main
file. The idea is to use the same `test.mk` file here and in fossa,
so any refactorings should keep that in mind.

For now v7 fails C++ compilation, disabling cxx dialect for now and fixing
in another PR.