Introduce feature selection machinery
Sets of features will be called "profiles" and we'll start with 3:
minimal, medium and full (the default).
Features are enabled and disabled by turning the knobs on and off.
Feature knobs follow common naming format: V7_ENABLE_Foo__bar (note the
double underscore), where "Foo" can be loosely defined as a "module" and
"bar" as a feature within it. "Feature" can be as small as a single
function or method or a collection of closely related functions/methods.
Slicing of the Date included in this commit illustrates intended use.

Full profile has all the features enabled automatically by grepping for
all the feature names and turning them on, other profiles will be
defined manually (current definitions are just skeletons).