Add a make variable that detects gcc vs clang
It was written to be able to disable some gcc specific warnings,
but it turns out that we don't want to disable them anymore.
However, it might be useful soon for other kind of compiler specific flags,
and it turned out to be a >15m task so I'd spare it to the next guy.