Fixed a couple of issues in String.prototype.split() :
First issue:

If pattern does not match an empty string, then the empty substring
at the end of the source string should be included in the resulting
array, but it wasn't. That is:

'123'.split(RegExp('.')) should return [ '', '', '', '' ],
but previously it returned [ '', '', '' ].

Second issue:

''.split(/.*/) should be [], not [ '' ], since given pattern
matches an empty string.

Details in spec:
http://www.ecma-international.org/ecma-262/5.1/#sec-15.5.4.14

Added a couple of new tests.